[ST7-feature/upload-file] - update and customize errors of multer middleware
