+++ stationery-api/config/multer-config.js
@@ -24,5 +24,6 @@ module.exports = {
   MULTIPLE_TYPE,
   SINGLE_UPLOAD_FIELD,
   MULTIPLE_UPLOAD_FIELD,
+  MAX_FILE_SIZE,
   MAX_FILES_ALLOWED,
 };
