+++ stationery-api/routes/upload-routes.js
@@ -1,10 +1,10 @@
 const { Router } = require("express");
 
-const dynamicUpload = require("../middlewares/multer-middleware");
+const multerMiddleware = require("../middlewares/multer-middleware");
 const uploadController = require("../controllers/upload-controller");
 
 const router = Router();
 
-router.post("", dynamicUpload, uploadController);
+router.post("", multerMiddleware, uploadController);
 
 module.exports = router;
